+++ packages/docs/src/stories/Avatar.stories.tsx
@@ -17,7 +17,7 @@ export default {
   },
 } as Meta<AvatarProps>
 
-export const Primary: StoryObj<AvatarProps> = {}
+export const Default: StoryObj<AvatarProps> = {}
 
 export const WithFallback: StoryObj<AvatarProps> = {
   args: {