+++ packages/docs/src/stories/TextInput.stories.tsx
@@ -1,10 +1,34 @@
 import type { StoryObj, Meta } from '@storybook/react'
-import { Box, Text, TextInput, TextInputProps } from '@skedle.me-ui/react'
+import {
+  Box,
+  IconButton,
+  Text,
+  TextInput,
+  TextInputProps,
+} from '@skedle.me-ui/react'
+import { Lock } from 'phosphor-react'
 
 export default {
-  title: 'Form/Text Input',
+  title: 'Form/TextInput',
   component: TextInput,
-  args: {},
+  args: {
+    placeholder: 'Type your name',
+    prefix: '',
+    disabled: false,
+    error: false,
+  },
+  argTypes: {
+    disabled: {
+      control: {
+        type: 'boolean',
+      },
+    },
+    error: {
+      control: {
+        type: 'boolean',
+      },
+    },
+  },
   decorators: [
     (Story) => {
       return (
@@ -12,7 +36,7 @@ export default {
           as="label"
           css={{ display: 'flex', flexDirection: 'column', gap: '$2' }}
         >
-          <Text size="sm">Email address</Text>
+          <Text size="sm">Name</Text>
           {Story()}
         </Box>
       )
@@ -20,11 +44,7 @@ export default {
   ],
 } as Meta<TextInputProps>
 
-export const Primary: StoryObj<TextInputProps> = {
-  args: {
-    placeholder: 'Type your name',
-  },
-}
+export const Default: StoryObj<TextInputProps> = {}
 
 export const Disabled: StoryObj<TextInputProps> = {
   args: {
@@ -32,9 +52,35 @@ export const Disabled: StoryObj<TextInputProps> = {
   },
 }
 
+export const WithError: StoryObj<TextInputProps> = {
+  args: {
+    error: true,
+  },
+}
+
 export const WithPrefix: StoryObj<TextInputProps> = {
   args: {
     prefix: 'skedle.me/',
     placeholder: 'your-username',
+  },
+}
+
+export const LeftButton: StoryObj<TextInputProps> = {
+  args: {
+    leftButton: (
+      <IconButton variant={'tertiary'} size={'sm'}>
+        <Lock weight="bold" />
+      </IconButton>
+    ),
+  },
+}
+
+export const RightButton: StoryObj<TextInputProps> = {
+  args: {
+    rightButton: (
+      <IconButton variant={'tertiary'} size={'sm'}>
+        <Lock weight="bold" />
+      </IconButton>
+    ),
   },
 }