+++ packages/tokens/src/colors.ts
@@ -3,20 +3,24 @@ export const colors = <const>{
   black: '#000',
 
   gray100: '#E1E1E6',
-  gray200: '#A9A9B2',
+  gray200: '#C4C4CC',
+  gray300: '#8D8D99',
   gray400: '#7C7C8A',
   gray500: '#505059',
   gray600: '#323238',
   gray700: '#29292E',
   gray800: '#202024',
   gray900: '#121214',
+  gray950: '#09090A',
 
-  yellow300: '#fcd34d',
+  yellow300: '#FCD34D',
   yellow400: '#FBBf24',
   yellow500: '#FFB900',
   yellow700: '#F59E0B',
 
   red600: '#FF5252',
+  red700: '#CC2937',
 
   green400: '#5BD76F',
+  green700: '#1B873F',
 }