+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @skedle.me-ui/tokens
 
+## 1.1.0
+
+### Minor Changes
+
+- Changes colors and font-sizes tokens
+
 ## 1.0.1
 
 ### Patch Changes